+++ test/test_qdmi.cpp
@@ -39,16 +39,14 @@ INSTANTIATE_TEST_SUITE_P(
     QDMIImplementationTest,
     // Test suite name
     // Parameters to test with
-    ::testing::Values(
-        std::tuple{"../examples/device/c/libc_device", "C",
-                   TEST_SESSION_MODE::READONLY},
-        std::tuple{"../examples/device/c/libc_device", "C",
-                   TEST_SESSION_MODE::READWRITE}
-        // std::tuple{"../examples/device/cxx/libcxx_device", "CXX",
-        //            TEST_SESSION_MODE::READONLY},
-        // std::tuple{"../examples/device/cxx/libcxx_device", "CXX",
-        //            TEST_SESSION_MODE::READWRITE}
-        ),
+    ::testing::Values(std::tuple{"../examples/device/c/libc_device", "C",
+                                 TEST_SESSION_MODE::READONLY},
+                      std::tuple{"../examples/device/c/libc_device", "C",
+                                 TEST_SESSION_MODE::READWRITE},
+                      std::tuple{"../examples/device/cxx/libcxx_device", "CXX",
+                                 TEST_SESSION_MODE::READONLY},
+                      std::tuple{"../examples/device/cxx/libcxx_device", "CXX",
+                                 TEST_SESSION_MODE::READWRITE}),
     [](const testing::TestParamInfo<
         std::tuple<std::string, std::string, TEST_SESSION_MODE>> &inf) {
       // Extract the last part of the file path
